- Add struct LogId which encloses the index and term of a very Log so that   comparing the order of uncommitted logs is ABA free now (which was not because   uncommitted logs might be truncated if leader changed and we doné¥æª compare   the term of the same index) - Move ConfigurationMananger from configure.h/cpp to configure_manager.h/cpp   and replace std::map with std::deque as its inner container - Unify LogManager::append_entries() for both followers and leader. Follower   append entries to LogStorage asynchronously as well. - Move conflicting checking a the appending entries from NodeImpl to   LogManager. - Remove LogManager::truncate_suffix. - Remove redundant LogManager::append() - Add flags é¥æaft_max_body_size (8M by default) which limits the size   of AppendEntriesRequest in a single RPC - Fix the bug that leader truncate_prefix races with append_entries. - SegmentedLogStorage::append_entries rejects non-continuous appending   entries. - Abort on storage error. - Add unit tests about appending conflicting entries - Add AsyncClosureGuard which starts a bthread to run the given Closure in   destructor
ISSUE=3196025

+++ src/raft/test/test_io_man.cpp
@@ -39,12 +39,12 @@ void* io_func(void* arg) {
         char buf[1024];
         snprintf(buf, sizeof(buf), "DATA:%d\n", i);
         int ret = raft::bthread_write(fd, buf, strlen(buf));
-        assert(ret == strlen(buf));
+        assert(ret == (int)strlen(buf));
 
         char buf2[1024];
         lseek(fd, old_off, SEEK_SET);
         ret = raft::bthread_read(fd, buf2, strlen(buf));
-        assert(ret == strlen(buf));
+        assert(ret == (int)strlen(buf));
         assert(0 == strncmp(buf, buf2, strlen(buf)));
     }
 
@@ -52,10 +52,10 @@ void* io_func(void* arg) {
     char buf[1024];
     snprintf(buf, sizeof(buf), "DATA END\n");
     int ret = raft::bthread_pwrite(fd, buf, strlen(buf), curr_off);
-    assert(ret == strlen(buf));
+    assert(ret == (int)strlen(buf));
     char buf2[1024];
     ret = raft::bthread_pread(fd, buf2, strlen(buf), curr_off);
-    assert(ret == strlen(buf));
+    assert(ret == (int)strlen(buf));
     assert(0 == strncmp(buf, buf2, strlen(buf)));
 
     ::close(fd);
